+++ core/src/test/kotlin/net/corda/core/transactions/LedgerTransactionQueryTests.kt
@@ -28,7 +28,7 @@ class LedgerTransactionQueryTests {
     @JvmField
     val testSerialization = SerializationEnvironmentRule()
     private val keyPair = generateKeyPair()
-    private val services = MockServices(rigorousMock<IdentityServiceInternal>().also {
+    private val services = MockServices(emptyList(), rigorousMock<IdentityServiceInternal>().also {
         doReturn(null).whenever(it).partyFromKey(keyPair.public)
     }, CordaX500Name("MegaCorp", "London", "GB"), keyPair)
     private val identity: Party = services.myInfo.singleIdentity()