+++ node/src/main/kotlin/net/corda/node/internal/protonwrapper/netty/SSLHelper.kt
@@ -1,0 +1,39 @@
+package net.corda.node.internal.protonwrapper.netty
+
+import io.netty.handler.ssl.SslHandler
+import net.corda.core.utilities.NetworkHostAndPort
+import net.corda.nodeapi.ArtemisTcpTransport
+import java.security.SecureRandom
+import javax.net.ssl.KeyManagerFactory
+import javax.net.ssl.SSLContext
+import javax.net.ssl.TrustManagerFactory
+
+internal fun createClientSslHelper(target: NetworkHostAndPort,
+                                   keyManagerFactory: KeyManagerFactory,
+                                   trustManagerFactory: TrustManagerFactory): SslHandler {
+    val sslContext = SSLContext.getInstance("TLS")
+    val keyManagers = keyManagerFactory.keyManagers
+    val trustManagers = trustManagerFactory.trustManagers
+    sslContext.init(keyManagers, trustManagers, SecureRandom())
+    val sslEngine = sslContext.createSSLEngine(target.host, target.port)
+    sslEngine.useClientMode = true
+    sslEngine.enabledProtocols = ArtemisTcpTransport.TLS_VERSIONS.toTypedArray()
+    sslEngine.enabledCipherSuites = ArtemisTcpTransport.CIPHER_SUITES.toTypedArray()
+    sslEngine.enableSessionCreation = true
+    return SslHandler(sslEngine)
+}
+
+internal fun createServerSslHelper(keyManagerFactory: KeyManagerFactory,
+                                   trustManagerFactory: TrustManagerFactory): SslHandler {
+    val sslContext = SSLContext.getInstance("TLS")
+    val keyManagers = keyManagerFactory.keyManagers
+    val trustManagers = trustManagerFactory.trustManagers
+    sslContext.init(keyManagers, trustManagers, SecureRandom())
+    val sslEngine = sslContext.createSSLEngine()
+    sslEngine.useClientMode = false
+    sslEngine.needClientAuth = true
+    sslEngine.enabledProtocols = ArtemisTcpTransport.TLS_VERSIONS.toTypedArray()
+    sslEngine.enabledCipherSuites = ArtemisTcpTransport.CIPHER_SUITES.toTypedArray()
+    sslEngine.enableSessionCreation = true
+    return SslHandler(sslEngine)
+}