+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/network/NetworkMapServer.kt
@@ -1,16 +1,12 @@
 package net.corda.testing.node.network
 
-import net.corda.core.crypto.Crypto
-import net.corda.core.crypto.SecureHash
-import net.corda.core.crypto.SignedData
-import net.corda.core.crypto.sha256
+import net.corda.core.crypto.*
 import net.corda.core.internal.cert
 import net.corda.core.internal.toX509CertHolder
 import net.corda.core.node.NodeInfo
 import net.corda.core.serialization.deserialize
 import net.corda.core.serialization.serialize
 import net.corda.core.utilities.NetworkHostAndPort
-import net.corda.core.utilities.hours
 import net.corda.nodeapi.internal.DigitalSignatureWithCert
 import net.corda.nodeapi.internal.NetworkMap
 import net.corda.nodeapi.internal.NetworkParameters
@@ -39,9 +35,11 @@ import javax.ws.rs.core.Response.ok
 
 class NetworkMapServer(cacheTimeout: Duration,
                        hostAndPort: NetworkHostAndPort,
+                       root_ca: CertificateAndKeyPair = ROOT_CA, // Default to ROOT_CA for testing.
                        vararg additionalServices: Any) : Closeable {
     companion object {
-        val stubNetworkParameter = NetworkParameters(1, emptyList(), 1.hours, 10, 10, Instant.now(), 10)
+        val stubNetworkParameter = NetworkParameters(1, emptyList(), 10485760, 40000, Instant.now(), 10)
+        private val serializedParameters = stubNetworkParameter.serialize()
 
         private fun networkMapKeyAndCert(rootCAKeyAndCert: CertificateAndKeyPair): CertificateAndKeyPair {
             val networkMapKey = Crypto.generateKeyPair(X509Utilities.DEFAULT_TLS_SIGNATURE_SCHEME)
@@ -56,9 +54,7 @@ class NetworkMapServer(cacheTimeout: Duration,
     }
 
     private val server: Server
-    // Default to ROOT_CA for testing.
-    // TODO: make this configurable?
-    private val service = InMemoryNetworkMapService(cacheTimeout, networkMapKeyAndCert(ROOT_CA))
+    private val service = InMemoryNetworkMapService(cacheTimeout, networkMapKeyAndCert(root_ca))
 
     init {
         server = Server(InetSocketAddress(hostAndPort.host, hostAndPort.port)).apply {
@@ -100,6 +96,11 @@ class NetworkMapServer(cacheTimeout: Duration,
     @Path("network-map")
     class InMemoryNetworkMapService(private val cacheTimeout: Duration, private val networkMapKeyAndCert: CertificateAndKeyPair) {
         private val nodeInfoMap = mutableMapOf<SecureHash, SignedData<NodeInfo>>()
+        private val parametersHash = serializedParameters.hash
+        private val signedParameters = SignedData(
+                serializedParameters,
+                DigitalSignature.WithKey(networkMapKeyAndCert.keyPair.public, Crypto.doSign(networkMapKeyAndCert.keyPair.private, serializedParameters.bytes))
+        )
 
         @POST
         @Path("publish")
@@ -115,7 +116,7 @@ class NetworkMapServer(cacheTimeout: Duration,
         @GET
         @Produces(MediaType.APPLICATION_OCTET_STREAM)
         fun getNetworkMap(): Response {
-            val networkMap = NetworkMap(nodeInfoMap.keys.map { it }, SecureHash.randomSHA256())
+            val networkMap = NetworkMap(nodeInfoMap.keys.map { it }, parametersHash)
             val serializedNetworkMap = networkMap.serialize()
             val signature = Crypto.doSign(networkMapKeyAndCert.keyPair.private, serializedNetworkMap.bytes)
             val signedNetworkMap = SignedNetworkMap(networkMap.serialize(), DigitalSignatureWithCert(networkMapKeyAndCert.certificate.cert, signature))
@@ -143,7 +144,7 @@ class NetworkMapServer(cacheTimeout: Duration,
         @Path("network-parameter/{var}")
         @Produces(MediaType.APPLICATION_OCTET_STREAM)
         fun getNetworkParameter(@PathParam("var") networkParameterHash: String): Response {
-            return Response.ok(stubNetworkParameter.serialize().bytes).build()
+            return Response.ok(signedParameters.serialize().bytes).build()
         }
 
         @GET